move cross-driver dependencies to extensions
