+++ drivers/hmis/app/models/hmis/hud/project.rb
@@ -20,9 +20,6 @@ class Hmis::Hud::Project < Hmis::Hud::Base
   has_many :project_cocs, **hmis_relation(:ProjectID, 'ProjectCoc'), inverse_of: :project, dependent: :destroy
   has_many :inventories, **hmis_relation(:ProjectID, 'Inventory'), inverse_of: :project, dependent: :destroy
   has_many :funders, **hmis_relation(:ProjectID, 'Funder'), inverse_of: :project, dependent: :destroy
-  has_many :external_ids, class_name: 'HmisExternalApis::ExternalId', as: :source
-  has_many :external_referral_requests, class_name: 'HmisExternalApis::ReferralRequest', dependent: :destroy
-  has_many :external_referral_postings, class_name: 'HmisExternalApis::ReferralPosting', dependent: :destroy
 
   has_and_belongs_to_many :project_groups,
                           class_name: 'GrdaWarehouse::ProjectGroup',
@@ -92,4 +89,6 @@ class Hmis::Hud::Project < Hmis::Hud::Base
     funders.where(end_date: nil).update_all(end_date: operating_end_date)
     inventories.where(inventory_end_date: nil).update_all(inventory_end_date: operating_end_date)
   end
+
+  include RailsDrivers::Extensions
 end
