+++ drivers/hmis_external_apis/extensions/grda_warehouse/remote_credential_extension.rb
@@ -1,0 +1,28 @@
+###
+# Copyright 2016 - 2023 Green River Data Analysis, LLC
+#
+# License detail: https://github.com/greenriver/hmis-warehouse/blob/production/LICENSE.md
+###
+
+module HmisExternalApis
+  module GrdaWarehouse
+    module RemoteCredentialExtension
+      extend ActiveSupport::Concern
+
+      included do
+        has_many :external_ids,
+                 class_name: 'HmisExternalApis::ExternalId',
+                 foreign_key: :remote_credential_id,
+                 dependent: :restrict_with_exception
+
+        def self.mci
+          where(slug: 'mci').first!
+        end
+
+        def self.mper
+          where(slug: 'mper').first!
+        end
+      end
+    end
+  end
+end
