+++ drivers/hmis_external_apis/extensions/hmis/hud/project_extension.rb
@@ -1,0 +1,21 @@
+###
+# Copyright 2016 - 2023 Green River Data Analysis, LLC
+#
+# License detail: https://github.com/greenriver/hmis-warehouse/blob/production/LICENSE.md
+###
+
+module HmisExternalApis
+  module Hmis
+    module Hud
+      module ProjectExtension
+        extend ActiveSupport::Concern
+
+        included do
+          has_many :external_ids, class_name: 'HmisExternalApis::ExternalId', as: :source
+          has_many :external_referral_requests, class_name: 'HmisExternalApis::ReferralRequest', dependent: :destroy
+          has_many :external_referral_postings, class_name: 'HmisExternalApis::ReferralPosting', dependent: :destroy
+        end
+      end
+    end
+  end
+end
